Move CRC function into separate module
